requirements.txt
redis==3.4.1
Django==3.0.4
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0